+++ OUDS/Core/Components/Sources/Dialogs/Alert/Internal/AlertMessage/AlertMessageContent.swift
@@ -23,6 +23,7 @@ struct AlertMessageContent: View {
     let description: TextualContent?
     let bulletList: [TextualContent]
     let link: OUDSAlertMessage.Link?
+    let onClose: (() -> Void)?
 
     @Environment(\.theme) private var theme
 
@@ -53,14 +54,22 @@ struct AlertMessageContent: View {
             }
             .accessibilityElement(children: .combine)
             .accessibilityLabel(accessibilityLabel)
+            .accessibilitySortPriority(OUDSAlertMessage.textsAccessibilityPriority)
+            #if canImport(UIKit)
+                .modifier(AlertMessageFKACustomActionsModifier(link: link, onClose: onClose))
+            #endif
 
             // Action
             if let link, self.link?.position == .bottom {
                 OUDSLink(text: link.text, size: .default, action: link.action)
+                    .accessibilitySortPriority(OUDSAlertMessage.actionLinkAccessibilityPriority)
             }
         }
         .padding(.vertical, theme.alert.spacePaddingBlock)
         .fixedSize(horizontal: false, vertical: true)
+        .onAppear {
+            VoiceOverUtils.announce(accessibilityLabel)
+        }
     }
 
     // MARK: - Helpers
@@ -88,6 +97,8 @@ struct AlertMessageContent: View {
             "core_alertMessage_warning_a11y".localized() + ","
         case .negative:
             "core_alertMessage_negative_a11y".localized() + ","
+        case .info:
+            "core_alertMessage_info_a11y".localized() + ","
         default:
             ""
         }