+++ OUDS/Core/Components/Sources/Navigations/ToolBar/OUDSToolBarItem.swift
@@ -141,8 +141,8 @@ public struct OUDSToolBarItem: View, Identifiable {
         /// **Warning: if OS is iOS 26+ / Liquid Glass, the label will not appear**
         ///
         ///  - Parameters:
-        ///     - label; The optional string label displayed near to the back indicator
-        ///     - accessibilityLabel: The accessibility label to describe the back action that could be overtied if needed,d efault set to *core_common_back*
+        ///     - label: The optional string label displayed near to the back indicator
+        ///     - accessibilityLabel: The accessibility label to describe the back action that could be overridden if needed, default set to *core_common_back*
         ///     - action: The action to do when clicked. If *nil* (default) the button is disabled. By default the dismiss is done after `action` is called..
         case back(label: String? = nil, accessibilityLabel: String = "core_common_back".localized(), action: (() -> Void)? = nil)