+++ OUDS/Core/Tokens/SemanticTokens/Sources/Values/FontSemanticTokens.swift
@@ -15,6 +15,8 @@
 // WARNING: Not synchronized anymore with the Figjam / Figma by developers team
 // Create an issue for update https://github.com/Orange-OpenSource/ouds-ios/issues/new?template=token_update.yml
 
+import CoreGraphics // Needed for XCFramework generation
+
 // swiftlint:disable missing_docs
 
 /// This is a group of semantic tokens for **font**.